Moved chain symbols to their own bucket and collection module.
